crewai/pyproject.toml
[project]
name = "crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["crewai[tools]>=0.152.0",
    "python-dotenv>=1.1.1",]
